# Conflicts:
#	src/main/java/uk/co/ramp/event/ContactEventProcessor.java
#	src/main/java/uk/co/ramp/people/Case.java
#	src/main/java/uk/co/ramp/people/Human.java
#	src/main/java/uk/co/ramp/people/PopulationGenerator.java
#	src/test/java/uk/co/ramp/event/ContactEventProcessorTest.java

+++ src/main/java/uk/co/ramp/io/types/CmptRecord.java
@@ -15,7 +15,7 @@ import uk.co.ramp.people.VirusStatus;
 @JsonPropertyOrder({"time", "s", "e", "a", "p", "sym", "sev", "r", "d"})
 public interface CmptRecord {
 
-  static CmptRecord of(int time, Map<VirusStatus, Integer> counts) {
+  static CmptRecord of(double time, Map<VirusStatus, Integer> counts) {
     return ImmutableCmptRecord.builder()
         .time(time)
         .s(counts.get(SUSCEPTIBLE))
@@ -29,7 +29,7 @@ public interface CmptRecord {
         .build();
   }
 
-  int time();
+  double time();
 
   int s();
